requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
newsapi-python>=0.2.7
yfinance>=0.2.40
resend>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
